updated some funcs

+++ srcs/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: khanadat <khanadat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/21 04:19:21 by khanadat          #+#    #+#             */
-/*   Updated: 2025/06/23 20:32:34 by khanadat         ###   ########.fr       */
+/*   Updated: 2025/09/09 04:05:33 by khanadat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 int	ft_strcmp(const char *s1, const char *s2)
 {
-	return (ft_strncmp(s1, s2, INT_MAX));
+	return (ft_strncmp(s1, s2, SIZE_MAX));
 }
